requirements.txt
flask==2.3.3
python-telegram-bot==13.15
requests==2.31.0
gunicorn==21.2.0